🚨 Extract `permissions` from `assets.root_paths`
TG-2031
