+++ src/youwol/backends/assets/routers/__init__.py
@@ -1,4 +1,5 @@
 # relative
 from .files import router as router_files
 from .images import router as router_images
+from .permissions import router as router_permissions
 from .raw import router as router_raw
